requirements.txt

fastapi==0.116.1
gspread==5.10.0
uvicorn==0.37.0
passlib==1.7.4
python-jose==3.3.0
pydantic==2.11.7
bcrypt==4.0.1
requests==2.32.4
starlette==0.47.2
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
cryptography==45.0.6
httpx==0.28.1
idna==3.10
pycparser==2.22
typing_extensions==4.15.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
